refactor(build): extract runtime constants to decouple page/content bundles from service graph
PokerChaseService's constructor references Dexie, all stream classes, and the
stats registry, so any value-import of the class (or the ./app barrel that
re-exports it) pulls the entire data-processing dependency graph into
whatever bundle imports it. web_accessible_resource.ts and content_script.ts
only needed the static POKER_CHASE_ORIGIN / POKER_CHASE_SERVICE_EVENT
constants, not the class itself.

Move those constants into src/constants/runtime.ts (no dependency on Dexie,
streams, or db). PokerChaseService now imports and re-exposes them as static
readonly properties for backward compatibility. Updated
web_accessible_resource.ts, content_script.ts, App.tsx, Hud.tsx, Popup.tsx,
GameTypeFilterSection.tsx, and types/messages.ts to import constants/types
directly instead of going through the service class or the ./app barrel.

Bundle sizes (esbuild output):
- dist/web_accessible_resource.js: 420,399 -> 14,474 bytes (-96.6%)
- dist/content_script.js:          633,855 -> 362,583 bytes (-42.8%, still
  includes React + MUI as expected)
